FIX to compile on ubuntu 25.04

+++ src/sofa/pointcloud/components/PointCloudRenderer.cpp
@@ -455,7 +455,7 @@ void PointCloudRenderer::doDrawVisual(const sofa::core::visual::VisualParams* vp
     // Here we have geometries to draw and a camera that look at it.
     // We first send the camera parameters to the gl rendering backend, then the geometries.
     auto c = l_camera->getPosition();
-    Eigen::Vector3f cam_pos {c[0],c[1],c[2]};
+    Eigen::Vector3f cam_pos {(float)c[0],(float)c[1],(float)c[2]};
 
     vparams->getModelViewMatrix(dviewmat.data());
     vparams->getProjectionMatrix(dprojmat.data());
